Bugfixes for SexBabesVR, added support for sister sites SinsVR and StasyQ VR
